requirements.txt
chardet==3.0.4
coreapi==2.3.3
Django==2.2.5
django-allauth==0.41.0
django-cors-headers==3.2.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
pylint==2.5.2
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.6
